moved CLI bundle to utilities
