+++ app/plugins/eu.esdihumboldt.hale.app.cli/META-INF/MANIFEST.MF
@@ -5,9 +5,9 @@ Bundle-SymbolicName: eu.esdihumboldt.hale.app.cli;singleton:=true
 Bundle-Version: 2.9.5.qualifier
 Bundle-Vendor: wetransform GmbH
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Import-Package: eu.esdihumboldt.hale.common.cli,
- eu.esdihumboldt.hale.common.cli.extension,
- eu.esdihumboldt.hale.common.cli.impl,
- eu.esdihumboldt.hale.common.core,
+Import-Package: eu.esdihumboldt.hale.common.core,
+ eu.esdihumboldt.util.cli,
+ eu.esdihumboldt.util.cli.extension,
+ eu.esdihumboldt.util.cli.impl,
  org.eclipse.equinox.app;version="1.1.0",
  org.osgi.framework;version="1.8.0"
