--- common/plugins/eu.esdihumboldt.hale.common.cli/src/eu/esdihumboldt/hale/common/cli/impl/ContextImpl.groovy
@@ -1,12 +1,0 @@
-package eu.esdihumboldt.hale.common.cli.impl
-
-import eu.esdihumboldt.hale.common.cli.CommandContext;
-import groovy.transform.Immutable
-
-@Immutable
-class ContextImpl implements CommandContext {
-
-  String baseCommand
-  String commandName
-
-}
