+++ ui/features/eu.esdihumboldt.hale.ui.feature/feature.xml
@@ -472,7 +472,7 @@
          unpack="false"/>
 
    <plugin
-         id="eu.esdihumboldt.hale.common.cli"
+         id="eu.esdihumboldt.util.cli"
          download-size="0"
          install-size="0"
          version="0.0.0"
