+++ util/plugins/eu.esdihumboldt.util.cli/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>eu.esdihumboldt.hale.common.cli</name>
+	<name>eu.esdihumboldt.util.cli</name>
 	<comment></comment>
 	<projects>
 	</projects>
