+++ util/plugins/eu.esdihumboldt.util.cli/.settings/org.eclipse.core.resources.prefs
