+++ util/plugins/eu.esdihumboldt.util.cli/build.properties
