+++ util/plugins/eu.esdihumboldt.util.cli/schema/eu.esdihumboldt.hale.cli.exsd
