+++ util/plugins/eu.esdihumboldt.util.cli/src/eu/esdihumboldt/util/cli/impl/ContextImpl.groovy
@@ -1,0 +1,11 @@
+package eu.esdihumboldt.util.cli.impl
+
+import eu.esdihumboldt.util.cli.CommandContext
+import groovy.transform.Immutable
+
+@Immutable
+class ContextImpl implements CommandContext {
+
+	String baseCommand
+	String commandName
+}
